+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/__init__.py
@@ -4,42 +4,6 @@
 # license information.
 # --------------------------------------------------------------------------
 
-from .._shared.policies_async import ExponentialRetry, LinearRetry
-from .._shared.models import(
-    LocationMode,
-    ResourceTypes,
-    AccountSasPermissions,
-    StorageErrorCode
-)
-from ..models import (
-    BlobType,
-    BlockState,
-    StandardBlobTier,
-    PremiumPageBlobTier,
-    SequenceNumberAction,
-    PublicAccess,
-    Logging,
-    Metrics,
-    RetentionPolicy,
-    StaticWebsite,
-    CorsRule,
-    ContainerProperties,
-    BlobProperties,
-    LeaseProperties,
-    ContentSettings,
-    CopyProperties,
-    BlobBlock,
-    PageRange,
-    AccessPolicy,
-    ContainerSasPermissions,
-    BlobSasPermissions,
-)
-from .models import (
-    ContainerPropertiesPaged,
-    BlobPropertiesPaged,
-    BlobPrefix
-)
-from .download_async import StorageStreamDownloader
 from .blob_client_async import BlobClient
 from .container_client_async import ContainerClient
 from .blob_service_client_async import BlobServiceClient
@@ -50,36 +14,5 @@ __all__ = [
     'BlobServiceClient',
     'ContainerClient',
     'BlobClient',
-    'BlobType',
     'LeaseClient',
-    'StorageErrorCode',
-    'ExponentialRetry',
-    'LinearRetry',
-    'LocationMode',
-    'BlockState',
-    'StandardBlobTier',
-    'PremiumPageBlobTier',
-    'SequenceNumberAction',
-    'PublicAccess',
-    'Logging',
-    'Metrics',
-    'RetentionPolicy',
-    'StaticWebsite',
-    'CorsRule',
-    'ContainerProperties',
-    'ContainerPropertiesPaged',
-    'BlobProperties',
-    'BlobPropertiesPaged',
-    'BlobPrefix',
-    'LeaseProperties',
-    'ContentSettings',
-    'CopyProperties',
-    'BlobBlock',
-    'PageRange',
-    'AccessPolicy',
-    'ContainerSasPermissions',
-    'BlobSasPermissions',
-    'ResourceTypes',
-    'AccountSasPermissions',
-    'StorageStreamDownloader',
 ]