+++ sdk/storage/azure-storage-queue/azure/storage/queue/__init__.py
@@ -25,14 +25,12 @@ from ._message_encoding import (
 from .models import (
     QueueMessage,
     QueueProperties,
-    QueuePropertiesPaged,
     QueueSasPermissions,
     AccessPolicy,
     Logging,
     Metrics,
     CorsRule,
     RetentionPolicy,
-    MessagesPaged,
 )
 
 __version__ = VERSION
@@ -48,7 +46,6 @@ __all__ = [
     'StorageErrorCode',
     'QueueMessage',
     'QueueProperties',
-    'QueuePropertiesPaged',
     'QueueSasPermissions',
     'AccessPolicy',
     'TextBase64EncodePolicy',
@@ -61,5 +58,4 @@ __all__ = [
     'Metrics',
     'CorsRule',
     'RetentionPolicy',
-    'MessagesPaged',
 ]