+++ sdk/storage/azure-storage-queue/azure/storage/queue/queue_client.py
@@ -198,7 +198,7 @@ class QueueClient(StorageAccountHostsMixin):
             start=None,  # type: Optional[Union[datetime, str]]
             policy_id=None,  # type: Optional[str]
             ip=None,  # type: Optional[str]
-            protocol=None  # type: Optional[str]
+            **kwargs  # type: Any
         ):  # type: (...) -> str
         """Generates a shared access signature for the queue.
 
@@ -234,9 +234,8 @@ class QueueClient(StorageAccountHostsMixin):
             or address range specified on the SAS token, the request is not authenticated.
             For example, specifying sip='168.1.5.65' or sip='168.1.5.60-168.1.5.70' on the SAS
             restricts the request to those IP addresses.
-        :param str protocol:
-            Specifies the protocol permitted for a request made. The default value
-            is https,http.
+        :keyword str protocol:
+            Specifies the protocol permitted for a request made. The default value is https.
         :return: A Shared Access Signature (sas) token.
         :rtype: str
 
@@ -249,6 +248,7 @@ class QueueClient(StorageAccountHostsMixin):
                 :dedent: 12
                 :caption: Generate a sas token.
         """
+        protocol = kwargs.pop('protocol', None)
         if not hasattr(self.credential, 'account_key') and not self.credential.account_key:
             raise ValueError("No account SAS key available.")
         sas = QueueSharedAccessSignature(