+++ sdk/storage/azure-storage-queue/tests/test_queue_samples_message_async.py
@@ -39,7 +39,7 @@ class TestMessageQueueSamples(AsyncQueueTestCase):
         try:
             # [START async_set_access_policy]
             # Create an access policy
-            from azure.storage.queue.aio import AccessPolicy, QueueSasPermissions
+            from azure.storage.queue import AccessPolicy, QueueSasPermissions
             access_policy = AccessPolicy()
             access_policy.start = datetime.utcnow() - timedelta(hours=1)
             access_policy.expiry = datetime.utcnow() + timedelta(hours=1)