+++ sdk/storage/azure-storage-queue/HISTORY.md
@@ -9,6 +9,8 @@ To use a queue_url, the method `from_queue_url` must be used.
 - `set_queue_access_policy` has required parameter `signed_identifiers`.
 - `NoRetry` policy has been removed. Use keyword argument `retry_total=0` for no retries.
 - `NoEncodePolicy` and `NoDecodePolicy` have been removed. Use `message_encode_policy=None` and `message_decode_policy=None`.
+- Removed types that were accidentally exposed from two modules. Only `QueueServiceClient` and `QueueClient`
+should be imported from azure.storage.queue.aio
 - Some parameters have become keyword only, rather than positional. Some examples include:
   - `loop`
   - `max_concurrency`
@@ -16,6 +18,7 @@ To use a queue_url, the method `from_queue_url` must be used.
   - `timeout` etc.
 - `QueueMessage` has had its parameters renamed from `insertion_time`, `time_next_visible`, `expiration_time`
 to `inserted_on`, `next_visible_on`, `expires_on`, respectively. 
+
 
  **New features**